# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedUsageResponseModel": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "ACCEPTED"
            ],
            "type": "string"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaSnapshotModel"
          },
          "units": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer"
          },
          "usageDate": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "decision",
          "usageDate",
          "units",
          "quota"
        ],
        "type": "object"
      },
...
      "ConcurrentRequestRetryExhaustedProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "CONCURRENT_REQUEST_RETRY_EXHAUSTED"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              503
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Concurrent request retry exhausted"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:concurrent-request-retry-exhausted"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "CreateUsageEventDto": {
        "additionalProperties": false,
        "properties": {
          "units": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "units"
        ],
        "type": "object"
      },
      "DailyUsageItemModel": {
        "additionalProperties": false,
        "properties": {
          "limitUnits": {
            "minimum": 1,
            "type": "integer"
          },
          "remainingUnits": {
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "usageDate": {
            "format": "date",
            "type": "string"
          },
          "usedUnits": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "usageDate",
          "usedUnits",
          "limitUnits",
          "remainingUnits",
          "updatedAt"
        ],
        "type": "object"
      },
      "DailyUsageResponseModel": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DailyUsageItemModel"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "DependencyUnavailableProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "DEPENDENCY_UNAVAILABLE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              503
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Dependency unavailable"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:dependency-unavailable"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "IdempotencyKeyReusedProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "IDEMPOTENCY_KEY_REUSED"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              409
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Idempotency key reused"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:idempotency-key-reused"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InsufficientScopeProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INSUFFICIENT_SCOPE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              403
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Insufficient scope"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:insufficient-scope"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InternalErrorProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              500
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Internal server error"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:internal-error"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "InvalidApiKeyProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "INVALID_API_KEY"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              401
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Invalid API key"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:invalid-api-key"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
...
      "QuotaExceededProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "QUOTA_EXCEEDED"
            ],
            "type": "string"
          },
          "decision": {
            "enum": [
              "QUOTA_EXCEEDED"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "quota": {
            "$ref": "#/components/schemas/QuotaSnapshotModel"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              429
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Daily quota exceeded"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:quota-exceeded"
            ],
            "format": "uri",
            "type": "string"
          },
          "units": {
            "maximum": 10000,
            "minimum": 1,
            "type": "integer"
          },
          "usageDate": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId",
          "eventId",
          "decision",
          "usageDate",
          "units",
          "quota"
        ],
        "type": "object"
      },
      "QuotaSnapshotModel": {
        "additionalProperties": false,
        "properties": {
          "limit": {
            "maximum": 1000000000,
            "minimum": 1,
            "type": "integer"
          },
          "remaining": {
            "maximum": 1000000000,
            "minimum": 0,
            "type": "integer"
          },
          "resetAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "limit",
          "remaining",
          "resetAt"
        ],
        "type": "object"
      },
...
      "UnsupportedMediaTypeProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "UNSUPPORTED_MEDIA_TYPE"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              415
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Unsupported media type"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:unsupported-media-type"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      },
      "ValidationErrorItem": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "type": "object"
      },
      "ValidationErrorProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_ERROR"
            ],
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            },
            "type": "array"
          },
          "requestId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              400
            ],
            "type": "integer"
          },
          "title": {
            "enum": [
              "Validation failed"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "urn:api-usage-quota-service:problem:validation-error"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "metricsBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "projectApiKeyBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      },
      "systemAdminBearer": {
        "bearerFormat": "opaque",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Tenant-isolated API usage metering and daily quota service.",
    "title": "API Usage Metering and Quota Service",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/usage-events": {
      "post": {
        "operationId": "UsageController.ingest",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsageEventDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedUsageResponseModel"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Quota-Limit": {
                "description": "Daily quota limit.",
                "required": true,
                "schema": {
                  "maximum": 1000000000,
                  "minimum": 1,
                  "type": "integer"
                }
              },
              "X-Quota-Remaining": {
                "description": "Remaining daily quota.",
                "required": true,
                "schema": {
                  "maximum": 1000000000,
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "X-Quota-Reset": {
                "description": "UTC quota reset time as Unix epoch seconds.",
                "required": true,
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorProblem"
                }
              }
            },
            "description": "Problem Details (400).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidApiKeyProblem"
                }
              }
            },
            "description": "Problem Details (401).",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "required": true,
                "schema": {
                  "enum": [
                    "Bearer"
                  ],
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientScopeProblem"
                }
              }
            },
            "description": "Problem Details (403).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/IdempotencyKeyReusedProblem"
                }
              }
            },
            "description": "Problem Details (409).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/UnsupportedMediaTypeProblem"
                }
              }
            },
            "description": "Problem Details (415).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/QuotaExceededProblem"
                }
              }
            },
            "description": "Problem Details (429).",
            "headers": {
              "X-Quota-Limit": {
                "description": "Daily quota limit.",
                "required": true,
                "schema": {
                  "maximum": 1000000000,
                  "minimum": 1,
                  "type": "integer"
                }
              },
              "X-Quota-Remaining": {
                "description": "Remaining daily quota.",
                "required": true,
                "schema": {
                  "maximum": 1000000000,
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "X-Quota-Reset": {
                "description": "UTC quota reset time as Unix epoch seconds.",
                "required": true,
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorProblem"
                }
              }
            },
            "description": "Unexpected internal server error.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/DependencyUnavailableProblem"
                    },
                    {
                      "$ref": "#/components/schemas/ConcurrentRequestRetryExhaustedProblem"
                    }
                  ]
                }
              }
            },
            "description": "Dependency unavailable or concurrency retry exhausted.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKeyBearer": []
          }
        ],
        "tags": [
          "usage"
        ]
      }
    },
    "/v1/usage/daily": {
      "get": {
        "operationId": "DailyUsageController.list",
        "parameters": [
          {
            "description": "Inclusive range end as one canonical UTC YYYY-MM-DD date. It must be on or after from; duplicate and unknown query fields are rejected; the inclusive range may span at most 90 days.",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Inclusive range start as one canonical UTC YYYY-MM-DD date. Duplicate and unknown query fields are rejected; the inclusive range may span at most 90 days.",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyUsageResponseModel"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorProblem"
                }
              }
            },
            "description": "Problem Details (400).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidApiKeyProblem"
                }
              }
            },
            "description": "Problem Details (401).",
            "headers": {
              "WWW-Authenticate": {
                "description": "Bearer authentication challenge.",
                "required": true,
                "schema": {
                  "enum": [
                    "Bearer"
                  ],
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientScopeProblem"
                }
              }
            },
            "description": "Problem Details (403).",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalErrorProblem"
                }
              }
            },
            "description": "Unexpected internal server error.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyUnavailableProblem"
                }
              }
            },
            "description": "Required dependency unavailable.",
            "headers": {
              "X-Request-Id": {
                "description": "Server-generated request identifier.",
                "required": true,
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKeyBearer": []
          }
        ],
        "tags": [
          "usage"
        ]
      }
    }
  },
  "servers": [],
...
}